+++ rollup.config.mjs
@@ -1,0 +1,26 @@
+import { uglify } from 'rollup-plugin-uglify';
+
+export default [
+
+	{
+		input: './src/tsl-textures.js',
+		output: {
+			file: './dist/tsl-textures.js',
+			format: 'es',
+			banner: '// TSL Textures\n',
+		},
+	},
+
+	{
+		input: './src/tsl-textures.js',
+		output: {
+			file: './dist/tsl-textures.min.js',
+			format: 'es',
+			banner: '// TSL Textures\n',
+		},
+		plugins: [
+			uglify()
+		]
+	},
+
+]